refactor: split bot invite and end shortcuts
